requirements.txt
# Core dependencies
streamlit>=1.30.0
streamlit-pdf>=0.1.0
python-telegram-bot>=20.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.1.0
pytz>=2023.3
pillow>=10.0.0
httpx>=0.25.0
structlog>=23.0.0
rapidfuzz>=3.0.0
packaging>=21.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
freezegun>=1.2.0
black>=23.0.0
mypy>=1.5.0
ruff>=0.1.0
